+++ BayesSurv/NEWS.md
@@ -3,13 +3,13 @@
 * continue to improve help files
 * fix hard-coding parts
 * add more functions for the visualization of results
-* add graphical learning for the 'Pooled' model?
 
 # BayesSurv 0.0-4
 
 * Bring back the original model in Madjar et al. (2021)
 * Extend Madjar et al. (2021) with fixed MRF graph for multiple subgroups
 * Add C++ functions `updateBH_cpp()`, `updateBH_list_cpp()`, `calJpost_helper_cpp`
+* Enable graphical learning for 'Pooled' model
 
 # BayesSurv 0.0-3